# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Online Application API",
        "description": "API Documentation for Online Application",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/getProducts/{family_id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "جلب المنتجات بتاعت فاميلي معينة",
                "operationId": "ff84e5a44ce519f295eaed60ed444197",
                "parameters": [
                    {
                        "name": "family_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 10
                    }
                ],
                "responses": {
                    "200": {
                        "description": "قائمة المنتجات",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "image": {
                                                        "type": "string",
                                                        "example": "http://example.com/storage/products/img.jpg",
                                                        "nullable": true
                                                    },
                                                    "product_id": {
                                                        "type": "integer",
                                                        "example": 101
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "منتج A"
                                                    },
                                                    "price": {
                                                        "type": "number",
                                                        "example": 57.5
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "enum": [
                                                            "in stock",
                                                            "out of stock"
                                                        ],
                                                        "example": "in stock"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/getProductDetails/{product_id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "تفاصيل منتج معين",
                "operationId": "da93e963727ef1a7c9e76c632a93fb03",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 101
                    }
                ],
                "responses": {
                    "200": {
                        "description": "تفاصيل المنتج",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "image": {
                                                    "type": "string",
                                                    "example": "http://example.com/storage/products/img.jpg",
                                                    "nullable": true
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "منتج A"
                                                },
                                                "code": {
                                                    "type": "integer",
                                                    "example": 101
                                                },
                                                "category": {
                                                    "type": "string",
                                                    "example": "مشروبات"
                                                },
                                                "price": {
                                                    "type": "number",
                                                    "example": 50
                                                },
                                                "tax": {
                                                    "type": "number",
                                                    "example": 7.5
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "enum": [
                                                        "in stock",
                                                        "out of stock"
                                                    ],
                                                    "example": "in stock"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "المنتج مش موجود",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "المنتج مش موجود"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}